+++ src/views/components/atoms/Form.tsx
@@ -1,0 +1,10 @@
+type FormProps = {
+  children: React.ReactNode;
+  [key: string]: any;
+};
+
+export const Form = (props: FormProps) => {
+  const { children, ...others } = props;
+
+  return <form {...others}>{children}</form>;
+};
